requirements.txt
mysql-connector-python
yahoo-fin         
yahooquery
yahoofinancials
pandas
